Split implementation into different files
