+++ cmd/new-api-example/crdt/crdt.go
@@ -533,278 +533,3 @@ loop:
 	}
 	return value, j - i, isDeleted
 }
-
-// ----
-
-type String struct {
-	tree   *CausalTree
-	atomID AtomID
-	minLoc int
-}
-
-func (*String) isValue() {}
-
-type Char struct {
-	tree   *CausalTree
-	atomID AtomID
-	minLoc int
-}
-
-func (s *String) Snapshot() string {
-	loc := s.tree.searchAtom(s.atomID, s.minLoc)
-	str, _, _ := s.tree.snapshotString(loc)
-	return str
-}
-
-func (s *String) StringCursor() *StringCursor {
-	return &StringCursor{s.tree, s.atomID, s.minLoc}
-}
-
-func (s *String) Cursor() Cursor {
-	return s.StringCursor()
-}
-
-// ----
-
-type StringCursor struct {
-	tree   *CausalTree
-	atomID AtomID
-	minLoc int
-}
-
-func (c *StringCursor) Index(i int) {
-	if i < 0 {
-		panic("Invalid negative index")
-	}
-	t := c.tree
-	loc := t.searchAtom(c.atomID, c.minLoc)
-	c.minLoc = loc
-
-	cnt := -1
-	j := loc + 1
-	for j < len(t.atoms) && t.withinBlock(j, loc) {
-		atom := t.atoms[j]
-		switch atom.tag {
-		case deleteTag:
-			// String is already deleted, but do nothing
-			j++
-		case charTag:
-			size, isDeleted := t.charBlock(j)
-			if !isDeleted {
-				cnt++
-				if cnt == i {
-					loc = j
-					break
-				}
-			}
-			j += size
-		default:
-			panic(fmt.Sprintf("string: unexpected tag: %v", atom.tag))
-		}
-	}
-	if cnt < i {
-		panic(fmt.Sprintf("string: index out of range: %d (size=%d)", i, cnt))
-	}
-	c.minLoc = loc
-	c.atomID = t.atoms[loc].id
-}
-
-func (c *StringCursor) Element() *Char {
-	loc := c.tree.searchAtom(c.atomID, c.minLoc)
-	c.minLoc = loc
-	return &Char{c.tree, c.atomID, loc}
-}
-
-func (c *StringCursor) Insert(ch rune) (AtomID, int) {
-	loc := c.tree.searchAtom(c.atomID, c.minLoc)
-	id, charLoc := c.tree.addAtom(c.atomID, loc, charTag, int32(ch))
-	c.atomID = id
-	c.minLoc = charLoc
-	return id, charLoc
-}
-
-func (c *StringCursor) Delete() {
-	loc := c.tree.deleteAtom(c.atomID, c.minLoc)
-	c.atomID, c.minLoc = c.tree.findNonDeletedCause(loc)
-}
-
-// ----
-
-type Counter struct {
-	tree   *CausalTree
-	atomID AtomID
-	minLoc int
-}
-
-func (*Counter) isValue() {}
-
-func (cnt *Counter) increment(x int32) {
-	loc := cnt.tree.searchAtom(cnt.atomID, cnt.minLoc)
-	cnt.minLoc = loc
-	cnt.tree.addAtom(cnt.atomID, loc, incrementTag, x)
-}
-
-func (cnt *Counter) Increment(x int32) { cnt.increment(+x) }
-func (cnt *Counter) Decrement(x int32) { cnt.increment(-x) }
-
-func (cnt *Counter) Snapshot() int32 {
-	loc := cnt.tree.searchAtom(cnt.atomID, cnt.minLoc)
-	x, _, _ := cnt.tree.snapshotCounter(loc)
-	return x
-}
-
-// ----
-
-type List struct {
-	tree   *CausalTree
-	atomID AtomID
-	minLoc int
-}
-
-func (*List) isValue() {}
-
-type Elem struct {
-	tree   *CausalTree
-	atomID AtomID
-	minLoc int
-}
-
-func (l *List) ListCursor() *ListCursor {
-	return &ListCursor{l.tree, l.atomID, l.minLoc}
-}
-
-func (l *List) Cursor() Cursor {
-	return l.ListCursor()
-}
-
-func (l *List) Snapshot() []interface{} {
-	loc := l.tree.searchAtom(l.atomID, l.minLoc)
-	xs, _, _ := l.tree.snapshotList(loc)
-	return xs
-}
-
-func (e *Elem) Clear() {
-	loc := e.tree.searchAtom(e.atomID, e.minLoc)
-	e.minLoc = loc
-
-	j := loc + 1
-	for j < len(e.tree.atoms) && e.tree.withinBlock(j, loc) {
-		atom := e.tree.atoms[j]
-		switch atom.tag {
-		case deleteTag:
-			// Elem is deleted, but do nothing.
-			j++
-		case stringTag, counterTag, listTag:
-			// Delete contents of elem.
-			e.tree.deleteAtom(atom.id, j)
-			return
-		default:
-			fmt.Println(e.tree.PrintTable())
-			panic(fmt.Sprintf("elem @ %d: unexpected tag @ %d: %v", loc, j, atom.tag))
-		}
-	}
-}
-
-func (e *Elem) SetString() *String {
-	loc := e.tree.searchAtom(e.atomID, e.minLoc)
-	e.minLoc = loc
-	return e.tree.newString(e.atomID, loc)
-}
-
-func (e *Elem) SetCounter() *Counter {
-	loc := e.tree.searchAtom(e.atomID, e.minLoc)
-	e.minLoc = loc
-	return e.tree.newCounter(e.atomID, loc)
-}
-
-func (e *Elem) SetList() *List {
-	loc := e.tree.searchAtom(e.atomID, e.minLoc)
-	e.minLoc = loc
-	return e.tree.newList(e.atomID, loc)
-}
-
-func (e *Elem) Value() Value {
-	loc := e.tree.searchAtom(e.atomID, e.minLoc)
-	e.minLoc = loc
-
-	j := loc + 1
-	for j < len(e.tree.atoms) && e.tree.withinBlock(j, loc) {
-		atom := e.tree.atoms[j]
-		switch atom.tag {
-		case deleteTag:
-			// Elem is deleted, but do nothing.
-			j++
-		case stringTag, counterTag, listTag:
-			return e.tree.valueOf(j)
-		default:
-			fmt.Println(e.tree.PrintTable())
-			panic(fmt.Sprintf("elem @ %d: unexpected tag @ %d: %v", loc, j, atom.tag))
-		}
-	}
-	// Empty elem.
-	return nil
-}
-
-// ----
-
-type ListCursor struct {
-	tree   *CausalTree
-	atomID AtomID
-	minLoc int
-}
-
-func (c *ListCursor) Index(i int) {
-	if i < 0 {
-		panic("Invalid negative index")
-	}
-	t := c.tree
-	loc := t.searchAtom(c.atomID, c.minLoc)
-	c.minLoc = loc
-
-	cnt := -1
-	j := loc + 1
-	for j < len(t.atoms) && t.withinBlock(j, loc) {
-		atom := t.atoms[j]
-		switch atom.tag {
-		case deleteTag:
-			// List is already deleted, but do nothing
-			j++
-		case elementTag:
-			size, isDeleted := t.elemBlock(j)
-			if !isDeleted {
-				cnt++
-				if cnt == i {
-					loc = j
-					break
-				}
-			}
-			j += size
-		default:
-			panic(fmt.Sprintf("list: unexpected tag: %v", atom.tag))
-		}
-	}
-	if cnt < i {
-		panic(fmt.Sprintf("list: index out of range: %d (size=%d)", i, cnt))
-	}
-	c.minLoc = loc
-	c.atomID = t.atoms[loc].id
-}
-
-func (c *ListCursor) Element() *Elem {
-	loc := c.tree.searchAtom(c.atomID, c.minLoc)
-	c.minLoc = loc
-	return &Elem{c.tree, c.atomID, loc}
-}
-
-func (c *ListCursor) Insert() *Elem {
-	loc := c.tree.searchAtom(c.atomID, c.minLoc)
-	id, charLoc := c.tree.addAtom(c.atomID, loc, elementTag, 0)
-	c.atomID = id
-	c.minLoc = charLoc
-	return &Elem{c.tree, id, charLoc}
-}
-
-func (c *ListCursor) Delete() {
-	loc := c.tree.deleteAtom(c.atomID, c.minLoc)
-	c.atomID, c.minLoc = c.tree.findNonDeletedCause(loc)
-}
